🚚 move project events in project module root
