+++ marketplace-domain/src/project/aggregate/mod.rs
@@ -1,6 +1,5 @@
 use crate::*;
 
-pub mod event;
 #[derive(Default, Debug, Clone, PartialEq, Eq)]
 pub struct Project;
