+++ marketplace-domain/src/project/mod.rs
@@ -2,7 +2,10 @@ mod projections;
 pub use projections::{LeadContributorProjection, MemberProjection};
 
 mod aggregate;
-pub use aggregate::{event::Event as ProjectEvent, Project as Aggregate};
+pub use aggregate::Project as Aggregate;
 
 mod projectors;
 pub use projectors::{LeadContributorProjector, MemberProjector};
+
+mod events;
+pub use events::Event;
